feat: 버스번호 자료형 변경

+++ modules/infrastructure/persistence/src/main/kotlin/com/moyeobus/infra/persistence/bus/entity/BusEntity.kt
@@ -17,7 +17,7 @@ class BusEntity (
 
     val operatorId: Long,
 
-    val busNumber: Int,
+    val busNumber: Long,
 
     val carNumber: String,
